filedb: Refactoring filedb -> database.
